requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-extensions==3.2.0
djangorestframework==3.13.1
drf-writable-nested==0.7.0
pytz==2022.1
sqlparse==0.4.2
